Decouple from dashboard
